internal/handler/handler_test.go
package handler_test

import (
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/A5TA/NutriTrack-Ai-backend/internal/handler"
	"github.com/gin-gonic/gin"
)

type testCase struct {
	name   string
	status int
}

func Test_PostMeal(t *testing.T) {
	testCases := []testCase{
		{name: "not implemented", status: http.StatusNotImplemented},
	}

	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			// Arrange
			r := gin.Default()
			r.POST("/", handler.PostMeal)

			req := httptest.NewRequest(http.MethodPost, "/", nil)
			w := httptest.NewRecorder()

			// Act
			r.ServeHTTP(w, req)

			// Assert
			if w.Code != tc.status {
				t.Errorf("Expected: %d, but got: %d", tc.status, w.Code)
			}
		})
	}
}

// Repeat similar patterns for other handlers
func Test_GetAllMeals(t *testing.T) {
	testCases := []testCase{
		{name: "not implemented", status: http.StatusNotImplemented},
	}

	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			// Arrange
			r := gin.Default()
			r.GET("/", handler.GetAllMeals)

			req := httptest.NewRequest(http.MethodGet, "/", nil)
			w := httptest.NewRecorder()

			// Act
			r.ServeHTTP(w, req)

			// Assert
			if w.Code != tc.status {
				t.Errorf("Expected: %d, but got: %d", tc.status, w.Code)
			}
		})
	}
}

// Similar test cases for `GetMeal`, `UpdateMeal`, and `DeleteMeal`
